# color-identifiers-mode.el: limit (looking-back) with line beginning